DEBUG: Add red background to test if container extends into safe area
This is a test commit to verify if the image container extends into the safe area.
The background is temporarily set to RED (#ff0000) so we can see it clearly.

If you see RED in the safe area: Container is positioned correctly, issue is with image
If you DON'T see RED in safe area: Container isn't extending properly

Please test and report back what you see.